Update return type

+++ app/include/app_sampler.h
@@ -12,7 +12,7 @@ void Sampler_join(void);
 
 // Getter
 int Sampler_getHistorySize(void);
-long Sampler_getDips(void);
+int Sampler_getDips(void);
 double Sampler_getAverageReading(void);
 long long Sampler_getNumSamplesTaken(void); 
 //double *Sampler_getHistory(int *size);
